requirements.txt
openai-whisper
openai
ffmpeg
fastapi
uvicorn
python-multipart
aiofiles